Change page size variables to 16-bit types.

+++ extEEPROM.cpp
@@ -29,7 +29,7 @@
  * nDevice is the number of EEPROM devices on the I2C bus.              *
  * pageSize is in bytes.                                                *
  *----------------------------------------------------------------------*/
-extEEPROM::extEEPROM(unsigned int deviceCapacity, byte nDevice, byte pageSize)
+extEEPROM::extEEPROM(unsigned int deviceCapacity, byte nDevice, unsigned int pageSize)
 {
     _dvcCapacity = deviceCapacity;
     _nDevice = nDevice;
@@ -53,7 +53,7 @@ byte extEEPROM::write(unsigned long addr, byte *values, byte nBytes)
     uint8_t deviceAddr;     //I2C device address
     uint8_t txStatus;       //transmit status
     uint8_t nWrite;         //number of bytes to write
-    uint8_t nPage;          //number of bytes remaining on current page, starting at addr
+    uint16_t nPage;         //number of bytes remaining on current page, starting at addr
     
     if (addr + nBytes > _totalCapacity) {   //will this write go the top of the EEPROM?
         return 9;                           //yes, tell the caller
@@ -100,7 +100,7 @@ byte extEEPROM::read(unsigned long addr, byte *values, byte nBytes)
     byte deviceAddr;
     byte rxStatus;
     uint8_t nRead;              //number of bytes to read
-    uint8_t nPage;              //number of bytes remaining on current page, starting at addr
+    uint16_t nPage;             //number of bytes remaining on current page, starting at addr
 
     if (addr + nBytes > _totalCapacity) {   //will this read take us past the top of the EEPROM?
         return 9;                           //yes, tell the caller
